fix: do not fail on unknown fields for SA config, modify env var prefix (#695)
* fix: do not fail on unknown fields for SA config, modify env var prefix

The rationale behind this is that certain settings might inject environment variables with the same prefix that we expect, so it will be considered for parsing and we had configured our input parsing to fail if there were unknown fields present.

This, then, disables that restriction so injected env vars can coexist with the config format we expect. We are also modifying the expected env var prefix to minimize chances of this happening anyways.

Collisions are not a problem since were possible with allowing env vars on the first place.

* test: fix for new prefix

* test: fix test for unknown fields

* test: fix integration

* test: env var-based tests should run sequentially or they'll collision