refactor: more robust health comparison (#757)
This makes the `PartialEq` implementations for `Healthy` and `Unhealthy` complain when additional fields are added to these structs.

Before, only an equality check was done on existing fields, but if new fields were added the compiler wouldn't complain, allowing a comparison to be inaccurate since the new fields were not accounted for.