refactor: merge hash and remote values management
Creates a `ConfigStatusManager` trait (with implementations for `onhost`/`k8s`) that is now in charge of managing both the hash repository and the yaml config. Previously, the management of these two items was split due to design decisions that are no longer in place

Given both the hash and the remote config values come in a single OpAMP payload, I think we should read and persist this data with a unique service, and only split if absolutely needed.