refactor: do not read env vars on the renderer
I noticed that with the (only) current implementation of `Renderer`, `TemplateRenderer`, if its `persister` is set to `None` then no I/O seems to happen **except** reading all the environment variables via `env::vars` (an operation that may also `panic`).

When testing with `TemplateRenderer` to render agent types, given the default value has the `persister` set to `None`, it is easy to achieve pure computations if we pass the env vars explicitly from outside or via a closure.

So, this PR:

- Modifies `retrieve_env_var_variables` so it does not panic.
- Changes `render`'s signature to accept the environment variables as a `HashMap<String, VariableDefinition>` parameter. This removes the need for setting env vars in unit tests and thus stop being `serial`.
  - Let me know if you would prefer the method to accept a `FnOnce() -> HashMap<String, VariableDefinition>` instead. I tried this but gave up because `mockall` does not play well with it and didn't want to remove the mock without checking with you first.

A next step for the renderer implementation to be completely pure could be modifying the `render` function return type to describe the effect we want to perform with the persister, if any, along the original return value, but given this impacts the behavior of the function in a major way I leave it for a future refactor.