feat: shutdown health checker gracefully (#1040)
We now wait for all the spawned threads of a given subagent to finish when we decide to stop the subagent.

We also included a refactor to wrap the threads with a new structure. This structure internally handles the creation
of the event publisher and consumer used to stop the thread from running infinitely. This mechanism is not
mandatory to use. However, it's the one in place in almost every thread spawned by the supervisors.
The advantage is that we no longer need to manually create the publisher, the consumer and take care of it before waiting the thread to finish. We can pass some metadata to the struct (e.g. agent id, thread name) and a function that we want to thread to execute and that's it. We can use the associated methods to start it.