chore: improve k8s supervisor logs (NR-370668) (#1245)
The health checker was printing the same warning multiple times in some circumstances. This didn't happen with healthy states, which we already handled in the past. If the previous and current states were healthy, we didn't log the new state. We extended that behaviour to unhealthy states. Now, states are only logged if they differ from the previous state.

Apart from that, we also took the opportunity to update some logs following the style guide.