build: address GoReleaser binary name limitation
Due to a GoReleaser limitation, the generated binary here must have the
same name as the final name we want to use in the generated and packaged
assets.

As our K8s build is more manual and does not have this problem, we can
name the on-host binary just `newrelic-agent-control` while the K8s one
is kept as `newrelic-agent-control-k8s`.