build: migrate config to GoReleaser v2 (#1533)
* build: use cargo-zigbuild

* chore: remove cross-rs references

* build: use GoReleaser v2 (cargo-zigbuild)

* build: address GoReleaser binary name limitation

Due to a GoReleaser limitation, the generated binary here must have the
same name as the final name we want to use in the generated and packaged
assets.

As our K8s build is more manual and does not have this problem, we can
name the on-host binary just `newrelic-agent-control` while the K8s one
is kept as `newrelic-agent-control-k8s`.

* chore: remove unneeded programs and scripts

* build: add .cargo/config.toml to manage crt-static for relevant targets

* build: increase ulimit

Addressing https://github.com/ziglang/zig/issues/23273

* Revert "build: increase ulimit"

This reverts commit d303a7fa0f6115144c7f6c4d3cd18a9ea64d93ad.