refactor: type-safe sub-agent ID
There are (or will be) functions around that are intended to operate with a SubAgentID and should not **ever** receive the AC AgentID or any other reserved AgentID that we add in the future. With this we can be a bit more sure that we are using the correct IDs.