fix(config_migrate): logs were disabled (#1792)
* fix(config_migrate): logs were disabled

We use a filtering mechanism in which we disabled all logs not coming from `newrelic_agent_control` crate, which excluded `newrelic_config_migrate`

the filters used make it tricky to configure logging properly for config_migrate. For example, debug log level cannot be activated, and previously the logs would be emitted depending on if the 'callsite' is inside the AC crate or not.

the error handling of `config_migrate` needs to be improved in any case.

* test: new default case